chore: refactored hs unit test
